+++ public/browser/chunk-C0iyj8vy.js
@@ -1,1 +1,0 @@
-import"./chunk-CD52DvrH.js";var t=[{path:``,loadComponent:()=>import(`./chunk-BSk8nrpc2.js`).then(o=>o.CallRecordsListComponent)},{path:`:id`,loadComponent:()=>import(`./chunk-CymNlXCH.js`).then(o=>o.CallRecordDetailComponent)}];export{t as CALL_ROUTES};