+++ public/browser/chunk-CvTXeTkn2.js
@@ -1,1 +1,0 @@
-import"./chunk-CD52DvrH.js";var t=[{path:``,loadComponent:()=>import(`./chunk-B9E11h6v2.js`).then(o=>o.AnalyticsPageComponent)}];export{t as ANALYTICS_ROUTES};