+++ public/browser/chunk-CwlyQfK7.js
@@ -1,0 +1,1 @@
+import"./chunk-CD52DvrH.js";var t=[{path:``,loadComponent:()=>import(`./chunk-DsXRw6bb2.js`).then(o=>o.CallRecordsListComponent)},{path:`:id`,loadComponent:()=>import(`./chunk-C5-nE4Fu.js`).then(o=>o.CallRecordDetailComponent)}];export{t as CALL_ROUTES};